Version is now 0.4.0.
This release is a major rewrite of the compass command line utility. It
is much more flexible and will better support other project types with
varying project structures.

There is a new configuration file format that allows you to specify
the sass, css, image and javascript locations for your project when
using compass in a stand-alone (or unsupported project) format.

Compass frameworks and plugins must now provide a manifest that
specifies what files and file types exist in the project template. This
paves the way for other types of templates as the same installation
framework that creates a project can now be used to install other
things.

The compass command line has better help text and more informative
output in general.