Following a similar API to how the Complier works, don't log anything (don't even load up the logger!) if we pass in the :quiet option.
I need this because of some conflicts with Compass::Logger, being referred
to as "Logger", which seemed to be pulling out the main Logger. This is a
seperate issue, but making this options[:quiet] API functional is a nice
work-around that keeps things from getting too messy.

Tests included, obviously.