Refactor hide-text
This fixes two major issues with the -9999px indent. With this method
the text is always flowing away from the container so there is no risk
of it flowing into view. Also it prevents a 9999px box from being
rendered by the browser saving tons of video memory and making things
faster ( see http://youtu.be/xuMWhto62Eo )