Allow arbitrary urls to be passed to replace-text.
There was no reason to restrict the input of replace-text to just paths
compatible with image-url, as such, this restriction is removed. URLs
can now be passed to replace-text. This makes it possible to use
external urls as well as to pass the result of image-url or inline-image
to the mixin.

However, the replace-text-with-dimensions does require a relative
path to the image so that the dimensions can be used, so we have added
an `$inline` argument for that mixin to support data-URIs properly.

Closes GH-893.