Replace broken Cornell link with Causes.com
On the Compass homepage, the link to Cornell University - CALS was
broken. I considered changing the link from
http://green.cals.cornell.edu to http://cals.cornell.edu, but that page
looked quite different than the thumbnail on the Compass homepage and I
was unable to verify that it actually used Compass. So instead, I
swapped it out for Causes.com, where we are making extensive use of
Compass.

Fixes #1576.